Added leaderboard

+++ src/main/java/egg/them/all/EggTimer.java
@@ -16,14 +16,14 @@ public class EggTimer {
     /*
         Gives all participants of EggTimer some amount of eggs. The exact calculation is pending.
     */
-    public void updateResoures(final Map<String, Integer> eggCount) {
+    public void updateResoures(final Map<String, Long> eggCount) {
         final long currentTime = System.currentTimeMillis();
         final int numEggBlessings = (int) (currentTime - lastUpdated) / updateFrequency * eggBlessingIncrement;
         final int eggBlessingAmount = numEggBlessings * eggBlessingIncrement;
 
         if (eggBlessingAmount > 0) {
             for (String user : eggCount.keySet()) {
-                int userEggCount = eggCount.get(user) + eggBlessingAmount;
+                long userEggCount = eggCount.get(user) + eggBlessingAmount;
                 eggCount.put(user, userEggCount);
             }
